+++ examples/openapi/app/source.ts
@@ -24,5 +24,4 @@ export const { getPage, getPages, pageTree } = loader({
 
 export const openapi = createOpenAPI({
   documentOrPath: path.resolve('./unkey.json'),
-  generateTypeScriptSchema: false,
 });