+++ packages/create-app/src/create-app.ts
@@ -1,5 +1,6 @@
 import path from 'node:path';
 import fs from 'node:fs/promises';
+import { tryGitInit } from '@/git';
 import localVersions from '../versions.json';
 import versionPkg from '../../create-app-versions/package.json';
 import type { PackageManager } from './auto-install';
@@ -12,11 +13,19 @@ export interface Options {
   outputDir: string;
   template: Template;
   tailwindcss: boolean;
-  installDeps: boolean;
   packageManager: PackageManager;
+
+  installDeps?: boolean;
+  initializeGit?: boolean;
+  log?: (message: string) => void;
 }
 
 export async function create(options: Options): Promise<void> {
+  const {
+    installDeps = true,
+    initializeGit = true,
+    log = console.log,
+  } = options;
   const projectName = path.basename(options.outputDir);
   const dest = path.resolve(cwd, options.outputDir);
   await copy(path.join(sourceDir, `template/+shared`), dest, (name) => {
@@ -29,9 +38,11 @@ export async function create(options: Options): Promise<void> {
   });
 
   await copy(path.join(sourceDir, `template/${options.template}`), dest);
+  log('Configured Typescript');
 
   if (options.tailwindcss) {
     await copy(path.join(sourceDir, `template/+tailwindcss`), dest);
+    log('Configured Tailwind CSS');
   }
 
   const packageJson = createPackageJson(projectName, options);
@@ -40,8 +51,13 @@ export async function create(options: Options): Promise<void> {
   const readMe = await getReadme(dest, projectName);
   await fs.writeFile(path.join(dest, 'README.md'), readMe);
 
-  if (options.installDeps) {
+  if (installDeps) {
     await autoInstall(options.packageManager, dest);
+    log('Installed dependencies');
+  }
+
+  if (initializeGit && tryGitInit(dest)) {
+    log('Initialized Git repository');
   }
 }